requirements-training.txt
# Requirements for training on Mac M3
# Install with: pip install -r requirements-training.txt

# Core training dependencies
torch>=2.1.0
transformers>=4.40.0
datasets>=2.14.0
accelerate>=0.27.0
peft>=0.9.0
bitsandbytes>=0.42.0
trl>=0.7.0

# Unsloth for optimized training on Mac Silicon (optional)
# If this fails, training will still work with standard transformers
# unsloth @ git+https://github.com/unslothai/unsloth.git

# Data processing
beautifulsoup4>=4.12.0
pandas>=2.0.0
tqdm>=4.66.0

# Embeddings for RAG
sentence-transformers>=2.3.0

# Vector database
chromadb>=0.4.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
